refactor!: create src/util/requests folder
