+++ src/events/mineflayer/chat/joinRequest.ts
@@ -1,9 +1,9 @@
 import { BlacklistEntry } from "../../../interfaces/BlacklistEntry";
 import { Event } from "../../../interfaces/Event";
 import _blacklist from "../../../util/_blacklist.json";
-import fetchHypixelPlayerProfile from "../../../util/fetchHypixelPlayerProfile";
-import fetchMojangProfile from "../../../util/fetchMojangProfile";
-import isFetchError from "../../../util/isFetchError";
+import fetchHypixelPlayerProfile from "../../../util/requests/fetchHypixelPlayerProfile";
+import fetchMojangProfile from "../../../util/requests/fetchMojangProfile";
+import isFetchError from "../../../util/requests/isFetchError";
 
 export default {
 	name: "chat:joinRequest",
