+++ src/util/requests/isFetchError.ts
@@ -1,4 +1,4 @@
-import { FetchError } from "../interfaces/FetchError";
+import { FetchError } from "../../interfaces/FetchError";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export default (response: any | FetchError): response is FetchError => {
